+++ lib/api/routes/topic.ts
@@ -568,14 +568,46 @@ export const topicRouter = createTRPCRouter({
     .input(z.string())
     .mutation(async ({ ctx, input }) => {
       try {
-        const data = await ctx.db.transaction(async () => {
-          await ctx.db
-            .delete(articleTopics)
-            .where(eq(articleTopics.topicId, input))
-          await ctx.db.delete(topics).where(eq(topics.id, input))
+        const topic = await ctx.db.query.topics.findFirst({
+          where: (topic, { eq }) => eq(topic.id, input),
         })
 
-        return data
+        if (topic) {
+          const checkIfTopicTranslationHasTopic =
+            await ctx.db.query.topicTranslations.findMany({
+              where: (topicTranslations, { eq }) =>
+                eq(topicTranslations.id, topic.topicTranslationId),
+              with: {
+                topics: true,
+              },
+            })
+
+          if (checkIfTopicTranslationHasTopic[0]?.topics.length === 1) {
+            const data = await ctx.db.transaction(async () => {
+              await ctx.db
+                .delete(articleTopics)
+                .where(eq(articleTopics.topicId, input))
+
+              await ctx.db.delete(topics).where(eq(topics.id, input))
+
+              await ctx.db
+                .delete(topicTranslations)
+                .where(eq(topicTranslations.id, topic.topicTranslationId))
+            })
+
+            return data
+          } else {
+            const data = await ctx.db.transaction(async () => {
+              await ctx.db
+                .delete(articleTopics)
+                .where(eq(articleTopics.topicId, input))
+
+              await ctx.db.delete(topics).where(eq(topics.id, input))
+            })
+
+            return data
+          }
+        }
       } catch (error) {
         console.error("Error:", error)
         if (error instanceof TRPCError) {
